Alternate get_access_token method for PmpAuth, Made use of it in Client
As per curl example given by Ryan Cavis, I have written an alternative method
for securing an access_token from the Homedoc. The previous method, still included,
followed the examples as given in the documentation and it only works for the
publish-pilot address. However, there is an alternate, simple construction which
works for the home doc, which means you can query the homedoc, get the access-token url
from it and request a token. Less fragile because you don't have to worry about
urls changing.

Also, added some tests to cover this new method (which pretty much just duplicated the
old tests).

Finally, edited the Client object to take advantage of this new method, so now I should
finally be able to give the client an endpoint, a client-id and client-secret and
it should be able to request its own tokens.

Next step: client tests.