Added collectiondoc files in order to begin working on collection doc.
Moved some of the methods that were previously in the client out to the navigabledoc
module and associated what had previously been `results` associated with the client
to the client as a NavigableDoc object.

Added method for creating queries, which will raise exceptions on
bad queries.

Working on a way to test the client using a testing server, based on
http.server from Python std lib. This will run in a separate process
while testing is run. Based on this, may be able to make some kind of content browser
using a small, local serve to serve the results.

Lots left to do, but the client is pretty workable at this point and can be
used to browse content.