Working on tests for pmp_client module. Discovered some bugs.
Bug in pager/navigabledoc interaction where multiplte 'navigation' elements are
present in a doc. Refined that a bit.

Also, found a bug in the history element of client: it was checking for
2 or more entries before using the history instead of greater than
zero.

NavigableDoc now uses its own links property in order to build nav
elements.

Built authdetails fixture for testing and fixed a bug in test-server
implementation as well.