Modified Client.gain_access to set Client.document and added unittest
The `gain_access` method of the pmp_client.Client object previously would
request the home-doc and then not save it as Client.document. However,
the expectation is that after 'gaining_access', a user should be able
to immediately begin querying. Modified method to set this attribute
as a NavigableDoc based on homedoc. Also, was able to pull options from
the document and to retrieve the href for credentials using the
NavigableDoc's methods, instead of `get_dict`, thereby eliminating an
import statement.

Added unittests for this behavior.

Also, bumped Changelog and version and edited TODO.