Fixed issue #3: test_pager.py failing
Due to the unordered nature of dictionaries, this issue
was only popping infrequently, but what happened was that
the Pager object was being passed all of a collection-docs
`links`, instead of the `navigation` attribute for its `links`.

As a result, it was sometimes finding links elsewhere in the doc
and it was also giving false-positives on whether or not
a document was `navigable` (the homedoc, for instance, was
always being set to `navigable`, when it's probably the only
doc that doesn't come over with navigation.

Added code to navigabledoc.py that sends only `navigation` links
to `Pager.update`, and updated `Pager.update` to check for None on
input.

Finally, edited the tests to reflect this change.