Added some error-handling in conn/auth for when responses are not good responses. This may cause more exception handling, but it's good to know at this point if requests are failing, which requests are failing and why.
Also, struggled for a bit with the filter_dict util. Couldn't decide
whether to make it recursively handle a heap of JSON data in order
to search out a particular object that contains the key/value pair.

That seems useful, so will return to it at some point, probably, and
I think that was my iniitial intention with that function. The trick
is how to write it.

Finally, finished tests for json_utils.