Added 'put' and 'delete' methods to the PmpConnector class.
Unfortunately, there's a lot of repeated code. I'm going to have to
think about the commonalities and abstract that stuff out because right now
the junk is just too ugly as it is.

Added a bunch of tests to test_conn.py, but there's a bunch of
repeated code in there too. It should definitely be simpler.

Added 'save' and 'delete' methods to the client, but these rely
pretty much on the PmpConnector's methods to achieve.

Adding collectiondoc/writeabledoc.py to the repo. This is a work
in progress, and I don't remember what I changed in navigabledoc.
I'm betting those are whitespace changes.