It should now be possible to edit a NavigableDoc by using the `edit` method and passing in 1) a list or tuple of keys/indices to access the value to be edited, and 2) a new value to set.
In addition, it is still totally possible to edit the `collectiondoc` attribute directly.
One issue that remains to be addressed is that if a NavigableDoc is to be submitted
to PMP, it must be verified in order to ensure that it does not include
read-only values ('created', 'modified' are the two big ones that jump out).

In addition, the NavigableDoc object has been given a `serialize` method, which
simply means a json.dump of the collectiondoc values associated with the object.

Questions remain to be answered, namely:

1) How to create a brand-new document?
2) How to verify a document to be sure it is valid before submitting?
3) How to take a profile sample and build a document from that?